+++ R/tidy_interpolator.R
@@ -438,6 +438,10 @@ create_meteo_interpolator <- function(meteo_with_topo, params = NULL, verbose = 
 #' @param filename file name for the interpolator nc file
 #' @param .overwrite logical indicating if the file should be overwritten if it
 #' already exists
+#' @param .verbose Logical indicating if the function must show messages and info.
+#' Default value checks \code{"meteoland_verbosity"} option and if not set, defaults
+#' to TRUE. It can be turned off for the function with FALSE, or session wide with
+#' \code{options(meteoland_verbosity = FALSE)}
 #' @return invisible interpolator object, to allow using this function as a
 #' step in a pipe
 #' @seealso Other interpolator functions: \code{\link{add_topo}()},
@@ -472,7 +476,11 @@ create_meteo_interpolator <- function(meteo_with_topo, params = NULL, verbose = 
 #' Victor Granda \enc{García}{Garcia}, EMF-CREAF
 #'
 #' @export write_interpolator
-write_interpolator <- function(interpolator, filename, .overwrite = FALSE) {
+write_interpolator <- function(
+  interpolator, filename,
+  .overwrite = FALSE,
+  .verbose = getOption("meteoland_verbosity", TRUE)
+) {
   # debug
   # browser()
 
@@ -524,7 +532,10 @@ write_interpolator <- function(interpolator, filename, .overwrite = FALSE) {
 
   ## write logic
   cf_conventions_web <- "https://cfconventions.org/cf-conventions/cf-conventions.html"
-  cli::cli_alert_info("Creating nc file following the NetCDF-CF conventions {.url {cf_conventions_web}}")
+  .verbosity_control(
+    cli::cli_alert_info("Creating nc file following the NetCDF-CF conventions {.url {cf_conventions_web}}"),
+    .verbose
+  )
 
   ### TODO add correct units
   list(prepared_data_list, names(prepared_data_list), prepared_data_units) |>
@@ -560,14 +571,20 @@ write_interpolator <- function(interpolator, filename, .overwrite = FALSE) {
       )
     )
 
-  cli::cli_alert_info("Adding spatial info to nc file")
+  .verbosity_control(
+    cli::cli_alert_info("Adding spatial info to nc file"),
+    .verbose
+  )
   filename <- ncdfgeom::write_geometry(
     nc_file = filename,
     geom_data = sf::st_as_sf(stars::st_get_dimension_values(interpolator, "station")),
     variables = c("instance_name", "time", "lat", "lon", names(prepared_data_list))
   )
 
-  cli::cli_alert_success("Done")
+  .verbosity_control(
+    cli::cli_alert_success("Done"),
+    .verbose
+  )
   return(invisible(interpolator))
 }
 
@@ -787,7 +804,7 @@ interpolator_calibration <- function(
     stations = NULL,
     update_interpolation_params = FALSE,
     variable = "MinTemperature",
-    N_seq = seq(5, 30, by = 5),
+    N_seq = seq(10, 30, by = 5),
     alpha_seq = seq(0.25, 10, by = 0.25),
     verbose = getOption("meteoland_verbosity", TRUE)
 ) {
@@ -817,6 +834,12 @@ interpolator_calibration <- function(
     is.numeric(alpha_seq),
     msg = "N_seq must be a numeric vector"
   )
+  if (variable %in% c("MinTemperature", "MaxTemperature")) {
+    assertthat::assert_that(
+      all(N_seq >= 10),
+      msg = "N_seq must start at 10 or bigger when calibrating temperature (Min or Max)"
+    )
+  }
 
   # get selected stations
   stations_sfc <- stars::st_get_dimension_values(interpolator, "station")
@@ -999,6 +1022,17 @@ interpolator_calibration <- function(
       if (variable %in% c("MinTemperature", "MaxTemperature", "DewTemperature")) {
         mae_matrix[as.character(i), as.character(j)] <-
           mean(abs(predicted_variable_matrix - selected_variable_matrix), na.rm = TRUE)
+
+        # check if NAs are generated (weights are zero for the parameter combination) and increase
+        # MAE to discard this parameter combination. Remember, increased MAE has to be bigger than
+        # the min_mae initial value (min_mae <- 9999999.0)
+        if (any(is.na(predicted_variable_matrix))) {
+          .verbosity_control(
+            cli::cli_ul("N = {.val {i}} & alpha = {.val {j}} removed from calibration as they generate zero weights in regression"),
+            verbose
+          )
+          mae_matrix[as.character(i), as.character(j)] <- 9999999.9
+        }
       } else {
         denominator_1 <- 0
         denominator_2 <- 0